gather refine
