--- oneflow/core/kernel/gather_kernel.h
@@ -1,23 +1,0 @@
-#ifndef ONEFLOW_CORE_KERNEL_GATHER_KERNEL_H_
-#define ONEFLOW_CORE_KERNEL_GATHER_KERNEL_H_
-
-#include "oneflow/core/kernel/kernel.h"
-
-namespace oneflow {
-
-template<DeviceType device_type, typename T>
-class GatherKernel final : public KernelIf<device_type> {
- public:
-  OF_DISALLOW_COPY_AND_MOVE(GatherKernel);
-  GatherKernel() = default;
-  ~GatherKernel() override = default;
-
- private:
-  const PbMessage& GetCustomizedOpConf() const override;
-  void ForwardDataContent(const KernelCtx& ctx,
-                          std::function<Blob*(const std::string&)> BnInOp2Blob) const override;
-};
-
-}  // namespace oneflow
-
-#endif  // ONEFLOW_CORE_KERNEL_GATHER_KERNEL_H_
